One solution to POJ1094 without using topological sort
Since there are only 26 letters in alphabet, the consumption of
adjacency matrix is tolerable. Every time we add an edge takes O(n)
time where n is the number of letters. When we add (x,y), if (y,x)
already exists, then inconsistency occurs. And if n*(n-1)/2 of the
matrix is filled with 1, then the sorted sequence is determined. And
the order is determined by the numbers of 1s in corresponding row for
each letter.